refactor: 컨벤션에 맞게 디렉터리 경로 수정
