+++ frontend/src/components/ChoiceTemplate/SampleImageList.tsx
@@ -1,7 +1,7 @@
 import { TEMPLATE_MENU } from '.';
-import { SampleCouponRes } from '../../../../types/api/response';
-import { SampleImage, SampleBackCouponImage } from '../../../../types/domain/coupon';
-import { TemplateMenu } from '../../../../types/utils';
+import { SampleCouponRes } from '../../types/api/response';
+import { SampleImage, SampleBackCouponImage } from '../../types/domain/coupon';
+import { TemplateMenu } from '../../types/utils';
 import { useSampleImages } from './hooks/useSampleImages';
 import { WarnMsg, SampleImageContainer, SampleImg, WarnMsgBox } from './style';
