+++ frontend/src/components/ChoiceTemplate/hooks/useSampleImages.tsx
@@ -1,8 +1,8 @@
 import { useLocation } from 'react-router-dom';
 import { useQuery } from '@tanstack/react-query';
-import { parseStampCount } from '../../../../../utils';
-import { getCouponSamples } from '../../../../../api/get';
-import { SampleCouponRes } from '../../../../../types/api/response';
+import { parseStampCount } from '../../../utils';
+import { getCouponSamples } from '../../../api/get';
+import { SampleCouponRes } from '../../../types/api/response';
 
 export const useSampleImages = () => {
   const location = useLocation();
