+++ frontend/src/components/ChoiceTemplate/style.tsx
