+++ frontend/src/components/CouponPreviewSection/index.tsx
@@ -1,7 +1,7 @@
-import { Spacing } from '../../../../style/layout/common';
+import { Spacing } from '../../style/layout/common';
 import FlippedCoupon, {
   FlippedCouponProps,
-} from '../../../Customer/CouponList/components/FlippedCoupon';
+} from '../../pages/Customer/CouponList/components/FlippedCoupon';
 import { CustomCouponLabel } from '../CustomCouponSection/style';
 
 type CouponPreviewSectionProps = FlippedCouponProps;
