+++ frontend/src/hooks/useMutateCouponPolicy.ts
@@ -1,7 +1,7 @@
 import { useMutation } from '@tanstack/react-query';
 import { useNavigate } from 'react-router-dom';
-import { postCouponSetting } from '../../../../api/post';
-import ROUTER_PATH from '../../../../constants/routerPath';
+import { postCouponSetting } from '../api/post';
+import ROUTER_PATH from '../constants/routerPath';
 
 export const useMutateCouponPolicy = () => {
   const navigate = useNavigate();
