+++ frontend/src/pages/Admin/CustomCouponDesign/components/StampCustomModal/index.tsx
